Fix: リファクタリング
