+++ src/types/index.d.ts
@@ -1,0 +1,10 @@
+export interface url_parameter {
+	from: string
+	to: string
+	y: string
+	m: string
+	d: string
+	hh: string
+	m2: string
+	m1: string
+}
